requirement.txt
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
discord-webhook==1.3.0
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
Faker==30.6.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
Markdown==3.7
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.20.2
